requirements.txt
bson==0.5.8
pymongo==3.7.1
click==8.0.4
Flask==2.0.3
itsdangerous==2.1.2
Jinja2==3.1.1
MarkupSafe==2.1.1
Werkzeug==2.0.3
azure-cosmos
numpy